pyproject.toml
[tool.poetry]
name = "sucinto"
version = "0.1.0"
description = ""
authors = ["Andre Fernandes <fpf.andre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.1"
openai = "^0.26.5"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
